pyproject.toml
[project]
name = "agent1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.4.1",
    "google-api-python-client>=2.189.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "ipykernel>=6.29.0",
    "langchain>=1.2.9",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.3.18",
    "langchain-core>=0.3.18",
    "langchain-groq>=0.2.0",
    "langchain-openai>=0.3.0",
    "langgraph>=1.0.0",
    "onnxruntime>=1.24.1",
    "openai>=2.16.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.0.1",
    "python-magic>=0.4.27",
    "sentence-transformers>=3.0.0",
]
